requirements.txt
# Core RAG Dependencies
flask==3.1.0
streamlit==1.41.0
chromadb==1.3.4
sentence-transformers>=2.3.0
watchdog==3.0.0
requests==2.31.0

# ML/AI
torch>=2.2.0
transformers>=4.35.0
anthropic>=0.44.0

# MCP (Model Context Protocol)
mcp>=0.9.0

# Additional utilities
python-dotenv==1.0.0
pyyaml==6.0.1
networkx>=3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-timeout>=2.1.0
httpx>=0.24.0

